refactor: service objects implementation
- 서비스 객체들의 인터페이스 분리
