+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/_core/security/JWTProvider.java
@@ -16,10 +16,10 @@ import java.util.Date;
 
 @Component
 public class JWTProvider {
-    // access-token expire time = 30 min
-    public final Long ACCESS_TOKEN_EXP = 1000L * 60 * 30;
-    // refresh-token expire time = 2 week
-    public final Long REFRESH_TOKEN_EXP = 1000L * 60 * 60 * 24 * 14;
+    // access-token expire time = 2 min
+    public final Long ACCESS_TOKEN_EXP = 1000L * 60 * 2;
+    // refresh-token expire time = 3 min
+    public final Long REFRESH_TOKEN_EXP = 1000L * 60 * 3;
     public final String TOKEN_PREFIX = "Bearer ";
     public final String AUTHORIZATION_HEADER = "Authorization";
     public final String REFRESH_HEADER = "Refresh";
