+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/match/MatchDTOConverter.java
@@ -1,11 +1,13 @@
 package com.kakao.sunsuwedding.match;
 
 import com.kakao.sunsuwedding._core.utils.DateFormat;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 
+@Component
 public class MatchDTOConverter {
-    public static List<MatchResponse.MatchDTO> toFindAllWithNoReviewDTO(List<Match> matches) {
+    public List<MatchResponse.MatchDTO> toFindAllWithNoReviewDTO(List<Match> matches) {
         // 플래너가 탈퇴한 경우는 조회 X
         List<Match> availableMatches = matches.stream().filter(match -> match.getPlanner() != null).toList();
