+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/PortfolioRestController.java
@@ -17,14 +17,14 @@ import java.util.List;
 @RequiredArgsConstructor
 @RequestMapping("/api/portfolios")
 public class PortfolioRestController {
-    private final PortfolioService portfolioService;
+    private final PortfolioServiceImpl portfolioServiceImpl;
 
     private static final int PAGE_SIZE = 10;
 
     @PostMapping(value = "")
     public ResponseEntity<?> addPortfolio(@RequestBody PortfolioRequest.AddDTO request,
                                            @AuthenticationPrincipal CustomUserDetails userDetails) {
-        portfolioService.addPortfolio(request, userDetails.getUser().getId());
+        portfolioServiceImpl.addPortfolio(request, userDetails.getUser().getId());
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
@@ -38,7 +38,7 @@ public class PortfolioRestController {
 
         Long userId = (userDetails == null) ? -1 : userDetails.getUser().getId();
         CursorRequest cursorRequest = new CursorRequest(cursor, PAGE_SIZE, name, location, minPrice, maxPrice);
-        PageCursor<List<PortfolioResponse.FindAllDTO>> response = portfolioService.findPortfolios(cursorRequest, userId);
+        PageCursor<List<PortfolioResponse.FindAllDTO>> response = portfolioServiceImpl.findPortfolios(cursorRequest, userId);
 
         return ResponseEntity.ok().body(ApiUtils.success(response));
     }
@@ -48,26 +48,26 @@ public class PortfolioRestController {
                                                   @AuthenticationPrincipal CustomUserDetails userDetails) {
 
         Long userId = (userDetails == null) ? -1 : userDetails.getUser().getId();
-        PortfolioResponse.FindByIdDTO portfolio = portfolioService.findPortfolioById(id, userId);
+        PortfolioResponse.FindByIdDTO portfolio = portfolioServiceImpl.findPortfolioById(id, userId);
         return ResponseEntity.ok().body(ApiUtils.success(portfolio));
     }
 
     @PutMapping(value = "")
     public ResponseEntity<?> updatePortfolio(@RequestBody PortfolioRequest.UpdateDTO request,
                                            @AuthenticationPrincipal CustomUserDetails userDetails) {
-        portfolioService.updatePortfolio(request, userDetails.getUser().getId());
+        portfolioServiceImpl.updatePortfolio(request, userDetails.getUser().getId());
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
     @DeleteMapping("")
     public ResponseEntity<?> deletePortfolio(@AuthenticationPrincipal CustomUserDetails userDetails) {
-        portfolioService.deletePortfolio(userDetails.getUser());
+        portfolioServiceImpl.deletePortfolio(userDetails.getUser());
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
     @GetMapping("/self")
     public ResponseEntity<?> myPortfolio(@AuthenticationPrincipal CustomUserDetails userDetails) {
-        PortfolioResponse.MyPortfolioDTO myPortfolio = portfolioService.myPortfolio(userDetails.getUser().getId());
+        PortfolioResponse.MyPortfolioDTO myPortfolio = portfolioServiceImpl.myPortfolio(userDetails.getUser().getId());
         return ResponseEntity.ok().body(ApiUtils.success(myPortfolio));
     }
 }
