+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/quotation/QuotationRestController.java
@@ -2,7 +2,7 @@ package com.kakao.sunsuwedding.quotation;
 
 import com.kakao.sunsuwedding._core.security.CustomUserDetails;
 import com.kakao.sunsuwedding._core.utils.ApiUtils;
-import com.kakao.sunsuwedding.match.MatchService;
+import com.kakao.sunsuwedding.match.MatchServiceImpl;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
@@ -14,27 +14,27 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/api/quotations")
 public class QuotationRestController {
-    private final QuotationService quotationService;
-    private final MatchService matchService;
+    private final QuotationServiceImpl quotationServiceImpl;
+    private final MatchServiceImpl matchServiceImpl;
 
     @PostMapping("")
     public ResponseEntity<?> addQuotation(@AuthenticationPrincipal CustomUserDetails userDetails,
                                              @RequestParam @Min(1) Long chatId,
                                              @Valid @RequestBody QuotationRequest.Add request) {
-        quotationService.addQuotation(userDetails.getUser(), chatId, request);
+        quotationServiceImpl.addQuotation(userDetails.getUser(), chatId, request);
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
     @GetMapping("")
     public ResponseEntity<?> findQuotationsByChatId(@RequestParam @Min(1) Long chatId) {
-        QuotationResponse.FindAllByMatchId response = quotationService.findQuotationsByChatId(chatId);
+        QuotationResponse.FindAllByMatchId response = quotationServiceImpl.findQuotationsByChatId(chatId);
         return ResponseEntity.ok().body(ApiUtils.success(response));
     }
 
     @GetMapping("/collect")
     public ResponseEntity<?> findQuotationsByUser(@AuthenticationPrincipal CustomUserDetails userDetails,
                                         @RequestParam(defaultValue = "0") Integer page) {
-        QuotationResponse.FindByUserDTO response = quotationService.findQuotationsByUser(userDetails.getUser(), page);
+        QuotationResponse.FindByUserDTO response = quotationServiceImpl.findQuotationsByUser(userDetails.getUser(), page);
         return ResponseEntity.ok().body(ApiUtils.success(response));
     }
 
@@ -42,7 +42,7 @@ public class QuotationRestController {
     public ResponseEntity<?> confirm(@AuthenticationPrincipal CustomUserDetails userDetails,
                                      @PathVariable @Min(1) Long quotationId,
                                      @RequestParam @Min(1) Long chatId) {
-        quotationService.confirm(userDetails.getUser(), chatId, quotationId);
+        quotationServiceImpl.confirm(userDetails.getUser(), chatId, quotationId);
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
@@ -51,14 +51,14 @@ public class QuotationRestController {
                                              @PathVariable @Min(1) Long quotationId,
                                              @RequestParam @Min(1) Long chatId,
                                              @Valid @RequestBody QuotationRequest.Update request) {
-        quotationService.updateQuotation(userDetails.getUser(), chatId, quotationId, request);
+        quotationServiceImpl.updateQuotation(userDetails.getUser(), chatId, quotationId, request);
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 
     @DeleteMapping("/{quotationId}")
     public ResponseEntity<?> deleteQuotation(@AuthenticationPrincipal CustomUserDetails userDetails,
                                              @PathVariable @Min(1) Long quotationId) {
-        quotationService.deleteQuotation(userDetails.getUser(), quotationId);
+        quotationServiceImpl.deleteQuotation(userDetails.getUser(), quotationId);
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 }
