+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/user/UserRestController.java
@@ -4,7 +4,7 @@ import com.kakao.sunsuwedding._core.security.CustomUserDetails;
 import com.kakao.sunsuwedding._core.security.JWTProvider;
 import com.kakao.sunsuwedding._core.utils.ApiUtils;
 import com.kakao.sunsuwedding.user.token.TokenDTO;
-import com.kakao.sunsuwedding.user.token.TokenService;
+import com.kakao.sunsuwedding.user.token.TokenServiceImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.util.Pair;
@@ -16,21 +16,21 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 @RequestMapping("/api/user")
 public class UserRestController {
-    private final UserService userService;
-    private final TokenService tokenService;
+    private final UserServiceImpl userServiceImpl;
+    private final TokenServiceImpl tokenServiceImpl;
     private final JWTProvider jwtProvider;
 
     // 회원가입
     @PostMapping("/signup")
     public ResponseEntity<?> signup(@RequestBody @Valid UserRequest.SignUpDTO requestDTO) {
-        UserResponse.FindUserId response = userService.signup(requestDTO);
+        UserResponse.FindUserId response = userServiceImpl.signup(requestDTO);
         return ResponseEntity.ok().body(ApiUtils.success(response));
     }
 
     // 로그인
     @PostMapping("/login")
     public ResponseEntity<?> login(@RequestBody @Valid UserRequest.LoginDTO requestDTO) {
-        Pair<TokenDTO, UserResponse.FindUserId> response = userService.login(requestDTO);
+        Pair<TokenDTO, UserResponse.FindUserId> response = userServiceImpl.login(requestDTO);
         return ResponseEntity
                 .ok()
                 .header(jwtProvider.AUTHORIZATION_HEADER, response.getFirst().accessToken())
@@ -40,7 +40,7 @@ public class UserRestController {
 
     @PostMapping("/token")
     public ResponseEntity<?> refresh(@AuthenticationPrincipal CustomUserDetails userDetails) {
-        TokenDTO tokens = tokenService.refreshAllTokens(userDetails.getUser());
+        TokenDTO tokens = tokenServiceImpl.refreshAllTokens(userDetails.getUser());
         return ResponseEntity
                 .ok()
                 .header(jwtProvider.AUTHORIZATION_HEADER, tokens.accessToken())
@@ -51,14 +51,14 @@ public class UserRestController {
     // 유저 정보 조회
     @GetMapping("/info")
     public ResponseEntity<?> getUserInfo(@AuthenticationPrincipal CustomUserDetails userDetails) {
-        UserResponse.FindById responseDTO = userService.findById(userDetails.getUser());
+        UserResponse.FindById responseDTO = userServiceImpl.findById(userDetails.getUser());
         return ResponseEntity.ok().body(ApiUtils.success(responseDTO));
     }
 
     // 회원 탈퇴
     @DeleteMapping("")
     public ResponseEntity<?> withdraw(@AuthenticationPrincipal CustomUserDetails userDetails) {
-        userService.withdraw(userDetails.getUser());
+        userServiceImpl.withdraw(userDetails.getUser());
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
 }
