+++ sunsu-wedding/src/test/java/com/kakao/sunsuwedding/quotation/QuotationRestControllerTest.java
@@ -5,7 +5,7 @@ import com.kakao.sunsuwedding._core.security.JWTProvider;
 import com.kakao.sunsuwedding._core.config.SecurityConfig;
 import com.kakao.sunsuwedding.user.UserRequest;
 import com.kakao.sunsuwedding.user.UserResponse;
-import com.kakao.sunsuwedding.user.UserService;
+import com.kakao.sunsuwedding.user.UserServiceImpl;
 import com.kakao.sunsuwedding.user.token.TokenDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -51,7 +51,7 @@ public class QuotationRestControllerTest {
     private ObjectMapper objectMapper;
 
     @Autowired
-    private UserService userService;
+    private UserServiceImpl userServiceImpl;
 
     @Autowired
     private JWTProvider jwtProvider;
@@ -61,7 +61,7 @@ public class QuotationRestControllerTest {
     @BeforeEach
     void beforeEach() {
         UserRequest.LoginDTO request = new UserRequest.LoginDTO("planner1@gmail.com", "planner1234!");
-        Pair<TokenDTO, UserResponse.FindUserId> response = userService.login(request);
+        Pair<TokenDTO, UserResponse.FindUserId> response = userServiceImpl.login(request);
         plannerToken = response.getFirst().accessToken();
     }
     // ============ 견적서 등록 테스트 ============
